# openapi.yml
openapi: 3.0.3
info:
  title: Bastion Core Banking & FinTech API
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    API Gateway reverse proxy untuk microservices Bastion:
    - **Identity Service** (:8081) -> `/v1/auth/*`
    - **Customer Service** (:8082) -> `/v1/customers/*`
    - **Wallet Service** (:8083) -> `/v1/wallets/*`
    
    *Catatan FinTech:* Semua nominal saldo (`balance`) menggunakan satuan **minor unit** (misal IDR Rp 10.000,00 = `1000000`).
servers:
  - url: http://localhost:8080
    description: Bastion API Gateway (Local Development)

tags:
  - name: 🔐 Identity & Auth
    description: Autentikasi, penerbitan JWT, dan manajemen sesi
  - name: 👤 Customer Profile
    description: Data nasabah dan identitas kepemilikan
  - name: 💳 Wallet & Balance
    description: Dompet multi-valas dan kontrol status saldo
  - name: 🩺 System Health
    description: Probe liveness & readiness gateway

paths:
  # =========================================================================
  # 🔐 IDENTITY & AUTH ENDPOINTS (/v1/auth)
  # =========================================================================
  /v1/auth/register:
    post:
      operationId: register
      tags:
        - 🔐 Identity & Auth
      summary: Registrasi Pengguna Baru
      description: Mendaftarkan akun baru, menyimpan password hash bcrypt, dan menerbitkan token sesi awal.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@bastion.id
                password:
                  type: string
                  minLength: 8
                  example: RahasiaBank123!
      responses:
        "201":
          description: Registrasi berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Payload tidak valid atau password terlalu pendek
        "409":
          description: Email sudah terdaftar sebelumnya

  /v1/auth/login:
    post:
      operationId: login
      tags:
        - 🔐 Identity & Auth
      summary: Login Pengguna
      description: "Verifikasi kredensial email & password. Jika 2FA aktif, mengembalikan temp_token (5 menit) dan flag two_factor_required: true."
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@bastion.id
                password:
                  type: string
                  example: RahasiaBank123!
      responses:
        "200":
          description: Login berhasil atau tantangan 2FA diberikan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: Email atau password salah

  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags:
        - 🔐 Identity & Auth
      summary: Silent Refresh Token
      description: Memperbarui Access Token yang kedaluwarsa menggunakan Refresh Token yang valid.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: Token berhasil diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                    example: 900
        "401":
          description: Refresh token kedaluwarsa atau tidak valid

  /v1/auth/logout:
    post:
      operationId: logout
      tags:
        - 🔐 Identity & Auth
      summary: Logout & Cabut Sesi
      security:
        - BearerAuth: []
      description: Memasukkan Access Token ke Redis blacklist dan mencabut session di database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Berhasil logout
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        "401":
          description: Token tidak valid

  /v1/auth/2fa/setup:
    post:
      operationId: setup2FA
      tags:
        - 🔐 Identity & Auth
      summary: Inisialisasi 2FA TOTP
      description: Menghasilkan shared secret Base32 baru dan otpauth:// URI untuk Google Authenticator. Secret disimpan terenkripsi AES-256-GCM.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Secret key dan URI berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorSetupResponse'
        "401":
          description: Belum terautentikasi

  /v1/auth/2fa/enable:
    post:
      operationId: enable2FA
      tags:
        - 🔐 Identity & Auth
      summary: Konfirmasi & Aktivasi 2FA
      description: "Memvalidasi 6-digit kode OTP dari aplikasi authenticator pengguna dan mengaktifkan two_factor_enabled: true."
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorEnableRequest'
      responses:
        "200":
          description: 2FA berhasil diaktifkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Two-factor authentication enabled successfully
        "400":
          description: Kode OTP tidak valid atau setup belum diinisialisasi
        "401":
          description: Belum terautentikasi

  /v1/auth/2fa/verify:
    post:
      operationId: verify2FA
      tags:
        - 🔐 Identity & Auth
      summary: Verifikasi Tantangan Login 2FA
      description: Memverifikasi temp_token dan 6-digit OTP saat login, lalu menerbitkan access & refresh token penuh.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorVerifyRequest'
      responses:
        "200":
          description: Verifikasi 2FA sukses, token sesi diterbitkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Format temp_token atau kode OTP tidak valid
        "401":
          description: Kode OTP salah atau temp_token telah kedaluwarsa

  /v1/auth/2fa/disable:
    post:
      operationId: disable2FA
      tags:
        - 🔐 Identity & Auth
      summary: Nonaktifkan 2FA
      description: Memerlukan verifikasi password akun sebelum menghapus secret TOTP dan menonaktifkan two_factor_enabled.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorDisableRequest'
      responses:
        "200":
          description: 2FA berhasil dinonaktifkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Two-factor authentication disabled successfully
        "401":
          description: Password akun salah atau sesi tidak valid

  # =========================================================================
  # 👤 CUSTOMER PROFILE ENDPOINTS (/v1/customers)
  # =========================================================================
  /v1/customers/me:
    get:
      operationId: getMyProfile
      tags:
        - 👤 Customer Profile
      summary: Ambil Profil Nasabah Saya
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Data profil nasabah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfile'
        "401":
          description: Belum terautentikasi
        "404":
          description: Profil customer belum dibuat

  # =========================================================================
  # 💳 WALLET ENDPOINTS (/v1/wallets)
  # =========================================================================
  /v1/wallets:
    get:
      operationId: listWallets
      tags:
        - 💳 Wallet & Balance
      summary: Ambil Daftar Dompet Nasabah
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Daftar seluruh dompet aktif nasabah
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
        "401":
          description: Tidak memiliki akses

    post:
      operationId: createWallet
      tags:
        - 💳 Wallet & Balance
      summary: Buat Dompet Valas Baru
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currency
              properties:
                currency:
                  type: string
                  enum: [IDR, USD]
                  example: USD
      responses:
        "201":
          description: Dompet berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        "400":
          description: Request body tidak valid
        "409":
          description: Dompet dengan mata uang ini sudah ada

  /v1/wallets/{id}:
    get:
      operationId: getWallet
      tags:
        - 💳 Wallet & Balance
      summary: Ambil Detail & Saldo Dompet Spesifik
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Detail dompet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        "401":
          description: Akses ditolak
        "404":
          description: Dompet tidak ditemukan

  /v1/wallets/{id}/freeze:
    post:
      operationId: freezeWallet
      tags:
        - 💳 Wallet & Balance
      summary: Bekukan Dompet (Emergency Kill-Switch)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dompet berhasil dibekukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        "401":
          description: Akses ditolak
        "404":
          description: Dompet tidak ditemukan

  /v1/wallets/{id}/unfreeze:
    post:
      operationId: unfreezeWallet
      tags:
        - 💳 Wallet & Balance
      summary: Buka Kembali Pembekuan Dompet
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dompet kembali aktif
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        "401":
          description: Akses ditolak
        "404":
          description: Dompet tidak ditemukan

  # =========================================================================
  # 🩺 SYSTEM HEALTH
  # =========================================================================
  /livez:
    get:
      operationId: livez
      tags:
        - 🩺 System Health
      summary: Liveness Probe
      security: []
      responses:
        "200":
          description: Gateway menyala dan menerima koneksi
        "500":
          description: Gateway mengalami kegagalan internal

  /readyz:
    get:
      operationId: readyz
      tags:
        - 🩺 System Health
      summary: Readiness Probe
      security: []
      responses:
        "200":
          description: Seluruh koneksi dependency siap melayani traffic
        "503":
          description: Layanan dependency belum siap

# ===========================================================================
# COMPONENTS / SCHEMAS
# ===========================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            status:
              type: string
              example: ACTIVE
            two_factor_enabled:
              type: boolean
              example: false
            created_at:
              type: string
              format: date-time
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900
        two_factor_required:
          type: boolean
          example: false
        temp_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    TwoFactorSetupResponse:
      type: object
      properties:
        secret:
          type: string
          example: JBSWY3DPEHPK3PXP
        otpauth_url:
          type: string
          example: otpauth://totp/Bastion:user@bastion.id?secret=JBSWY3DPEHPK3PXP&issuer=Bastion

    TwoFactorEnableRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
          example: "123456"

    TwoFactorVerifyRequest:
      type: object
      required:
        - temp_token
        - code
      properties:
        temp_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        code:
          type: string
          pattern: '^[0-9]{6}$'
          example: "123456"

    TwoFactorDisableRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          example: RahasiaBank123!

    CustomerProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        first_name:
          type: string
          example: Agam
        last_name:
          type: string
          example: Latiff
        phone_number:
          type: string
          example: "+6281234567890"
        status:
          type: string
          example: VERIFIED

    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        currency:
          type: string
          example: IDR
        balance:
          type: integer
          format: int64
          description: Minor units (misal 24500000 = Rp 245.000,00)
          example: 24500000
        max_balance_limit:
          type: integer
          format: int64
          example: 10000000000
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
          example: ACTIVE
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time